requirements.txt
requests>=2.31.0
python-telegram-bot>=20.7
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
APScheduler>=3.10.4
tabulate>=0.9.0
